import ept files; move import trajectory to a generic module
