+++ setup.py
@@ -72,7 +72,8 @@ setup(
             'import-ori = cli_li3ds.import_ori:ImportOri',
             'import-orimatis = cli_li3ds.import_orimatis:ImportOrimatis',
             'import-image = cli_li3ds.import_image:ImportImage',
-            'import-trajectory = cli_li3ds.import_trajectory:ImportTrajectory',
+            'import-trajectory = cli_li3ds.import_pointcloud:ImportTrajectory',
+            'import-ept = cli_li3ds.import_pointcloud:ImportEpt',
         ]
     }
 )
